+++ src/django/VLE/models.py
@@ -4,7 +4,6 @@ models.py.
 Database file
 """
 import os
-import re
 
 from django.contrib.auth.models import AbstractUser
 from django.core.exceptions import ValidationError
@@ -18,20 +17,7 @@ from VLE.utils.error_handling import (VLEParticipationError,
                                       VLEPermissionError, VLEProgrammingError,
                                       VLEUnverifiedEmailError)
 from VLE.utils.file_handling import get_path
-
-
-# from https://stackoverflow.com/q/25121165/5173684
-def strip_script_tags(page_source):
-    if page_source is None or page_source == '':
-        return page_source
-    result = page_source
-    pattern = re.compile(r'\s?on\w+="[^"]+"\s?')
-    result = re.sub(pattern, "", result)
-    pattern = re.compile(r"\s?on\w+='[^']+'\s?")
-    result = re.sub(pattern, "", result)
-    pattern = re.compile(r'<script[\s\S]+?/script>')
-    result = re.sub(pattern, "", result)
-    return result
+from VLE.utils import sanitization
 
 
 class Instance(models.Model):
@@ -443,7 +429,7 @@ class Assignment(models.Model):
         return self.due_date and self.due_date < now()
 
     def save(self, *args, **kwargs):
-        self.description = strip_script_tags(self.description)
+        self.description = sanitization.strip_script_tags(self.description)
 
         return super(Assignment, self).save(*args, **kwargs)
 
@@ -790,7 +776,7 @@ class Content(models.Model):
     )
 
     def save(self, *args, **kwargs):
-        self.data = strip_script_tags(self.data)
+        self.data = sanitization.strip_script_tags(self.data)
 
         return super(Content, self).save(*args, **kwargs)
 
@@ -822,7 +808,7 @@ class Comment(models.Model):
         if not self.pk:
             self.creation_date = timezone.now()
         self.last_edited = timezone.now()
-        self.text = strip_script_tags(self.text)
+        self.text = sanitization.strip_script_tags(self.text)
         return super(Comment, self).save(*args, **kwargs)
